changed the selection trigger to anytime on close
