+++ src/main/kotlin/usecases/LanguageToString.kt
@@ -1,0 +1,8 @@
+package usecases
+
+import data.Language
+
+fun Language.toString() : String {
+    return "${this.slug} (${this.name})"
+}
+
