+++ src/main/kotlin/widgets/LanguageSelection.kt
@@ -1,12 +1,11 @@
 package widgets
 
 import data.Language
+import javafx.application.Platform
 import javafx.beans.property.SimpleObjectProperty
 import javafx.collections.ObservableList
-import javafx.event.EventHandler
 import javafx.geometry.Insets
 import javafx.geometry.Pos
-import javafx.scene.Node
 import javafx.scene.control.Button
 import javafx.scene.control.ComboBox
 import javafx.scene.control.Label
@@ -18,12 +17,9 @@ import javafx.scene.layout.StackPane
 import javafx.scene.paint.Color
 import javafx.scene.paint.Paint
 import javafx.scene.shape.Rectangle
+import javafx.util.StringConverter
 import recources.UIColors
 import tornadofx.*
-import usecases.LanguageStringConverter
-
-
-
 
 /**
  * This class is used to make the drop-downs for adding new
@@ -50,13 +46,9 @@ class LanguageSelection(languages : ObservableList<Language>,
                         private val selectedLanguages : ObservableList<Language>
 ) : Fragment() {
 
-    //private val languageValues = languages.values.toMutableList()
-    //private val languageKeys = languages.keys.toMutableList()
-
     private val fp = flowpane {
 
         children.bind(selectedLanguages, ::addTagNode)
-
         vgrow = Priority.ALWAYS
 
         hgap = 6.0
@@ -77,7 +69,8 @@ class LanguageSelection(languages : ObservableList<Language>,
             converter = LanguageStringConverter()
 
             /**
-             * Allow filtered searching
+             * Allow filtered searching and filter based on a language's name,
+             * slug, and anglicized name
              */
             isEditable = true
             promptText = hint
@@ -90,28 +83,32 @@ class LanguageSelection(languages : ObservableList<Language>,
                 }.sorted()
             }
 
-            // force the 'input' to be what is in the text field over what is focused in the dropdown (old)
-            // this.editor.textProperty().addListener({ inputForcer, oldText, newText -> this.setValue(newText)})
-
-            // TODO: Find out how to shorten lambda syntax
-            /*
-            this.editor.onKeyPressed = EventHandler { event ->
-                println("here")
-                if (event.getCode().equals(KeyCode.ENTER)) {
-                    this.editor.text = languageToString(input.value)
+            /**
+             * Remove any text in the textfield when it is refocused
+             */
+            editor.focusedProperty().addListener {
+                obs, old, new -> run {
+                    Platform.runLater {
+                        if (editor.isFocused && !editor.text.isEmpty()) {
+                            editor.selectAll()
+                        }
+                    }
                 }
             }
-            this.setOnKeyPressed {  }
-            */
-            addEventFilter(ComboBox.ON_HIDDEN, {
+
+            /**
+             * make a selectable chip button when the dropdown closes and a valid language
+             * is selected
+             */
+            addEventFilter(ComboBox.ON_HIDDEN) {
                 println(input)
                 println(this.editor.text)
-                if (languages.map { languageToString(it) }.contains(this.editor.text)) {
+                //if (/*languages.map { languageToString(it) }.contains(this.editor.text)*/) {
                     if (languages.contains(input.value) && !selectedLanguages.contains(input.value)) {
                         selectedLanguages.add(0, input.value)
                     }
-                }
-            })
+                //}
+            }
 
         }
 
@@ -125,11 +122,11 @@ class LanguageSelection(languages : ObservableList<Language>,
     }
 
     // Given by the awesome Carl
-    private fun addTagNode(language : Language) : Node {
+    private fun addTagNode(language : Language) : StackPane {
 
         val background = Rectangle()
         background.fill = Paint.valueOf(UIColors.UI_PRIMARY)
-        //TODO: will need to make the dimensions dynamic
+        // dynamic sizing?
         background.arcWidth = 20.0
         background.arcHeight = 20.0
         background.width = 200.0
@@ -138,9 +135,12 @@ class LanguageSelection(languages : ObservableList<Language>,
 
         // dynamic padding?
         val label = Label(languageToString(language))
+        label.text
         val labelHBox = HBox(label)
         labelHBox.alignment = Pos.CENTER
         labelHBox.padding = Insets(20.0)
+        labelHBox.isPickOnBounds = false
+
 
         val deleteButton = Button("X")
         deleteButton.userData = language // pro tip (thanks Carl)
@@ -156,6 +156,7 @@ class LanguageSelection(languages : ObservableList<Language>,
         val labelDelB = HBox(deleteButton)
         labelDelB.alignment = Pos.CENTER_RIGHT
         labelDelB.padding = Insets(10.0)
+        labelDelB.isPickOnBounds = false
 
         //background.width = label.prefWidth + labelDelB.prefWidth
 
@@ -169,8 +170,7 @@ class LanguageSelection(languages : ObservableList<Language>,
         }
 
         // Find out how to shorten lambda syntax
-        tag.addEventFilter(MouseEvent.MOUSE_CLICKED, EventHandler<MouseEvent> { mouseEvent -> newSelected(language)})
-
+        tag.addEventFilter(MouseEvent.MOUSE_CLICKED) { newSelected(language) }
 
         return tag
     }
@@ -181,6 +181,7 @@ class LanguageSelection(languages : ObservableList<Language>,
      */
     private fun resetSelected() {
         val firstTag = fp.children[0]
+
         for (nodeOut in firstTag.getChildList().orEmpty()) {
             if (nodeOut is Rectangle) {
                 nodeOut.fill = Paint.valueOf(UIColors.UI_PRIMARY)
@@ -199,11 +200,11 @@ class LanguageSelection(languages : ObservableList<Language>,
         // for all objects in the flowpane
         for (children in elements) {
 
-            // for all children in each flowplane object
+            // for all children in each flowpane object
             for (nodeOut in children.getChildList().orEmpty()) {
 
                 // change the rectangle color
-                // and kep a reference
+                // and keep a reference
                 if (nodeOut is Rectangle) {
                     rectangleReference = nodeOut
                     nodeOut.fill = Color.WHITE
@@ -224,8 +225,28 @@ class LanguageSelection(languages : ObservableList<Language>,
         }
     }
 
+    /**
+     * This function is used as a toString() for a language object
+     */
     private fun languageToString(language : Language) : String {
         return "${language.slug} (${language.name})"
     }
 
+}
+
+
+/**
+ * This converter is used for the comboBox in order to display language objects as strings
+ * in the dropdown
+ */
+class LanguageStringConverter : StringConverter<Language>() {
+    private val mapLanguage = mutableMapOf<String, Language>()
+
+    override fun fromString(string: String?): Language? = string?.let { mapLanguage[it] }
+
+    override fun toString(language : Language?): String? {
+        val output = "${language?.slug} (${language?.name})"
+        if (language != null && !mapLanguage.containsKey(output)) mapLanguage[output] = language
+        return output
+    }
 }
